+++ src/main/java/de/schaack/ml/basics/BasicsApplication.java
@@ -1,43 +1,27 @@
 package de.schaack.ml.basics;
 
-import java.util.Date;
 import java.util.List;
 import java.util.logging.Logger;
 
 import de.schaack.ml.basics.config.logging.BufferingLogHandler;
 import de.schaack.ml.basics.config.logging.LoggerConfig;
 import de.schaack.ml.basics.config.logging.LoggerUtil;
-import de.schaack.ml.basics.config.properties.ConfigReader;
 
 public class BasicsApplication {
-	private static final Logger log = LoggerUtil.getLogger(BasicsApplication.class.getName());
+
+	private static Logger log = LoggerUtil.getLogger(BasicsApplication.class.getName());
 
 	public static void main(String... args) {
 		run(args);
 	}
 
 	private static void run(String[] args) {
-		// Initialization code
-		init();
-
 		try {
 			// Main logic
 			mainLogic(args);
 		} finally {
 			// Post-execution code
 			post();
-		}
-	}
-
-	private static void init() {
-		// Reference to ConfigReader will force its initialization
-		System.out.println(new Date() + " Initializing application...");
-
-		if (ConfigReader.getProperties() != null) {
-			System.out.println(new Date() + " Configuration files loaded successfully.");
-			ConfigReader.getProperties().forEach((key, value) -> log.config(key + "=" + value));
-		} else {
-			System.err.println(new Date() + " Failed to load configuration files.");
 		}
 	}