+++ src/main/java/de/schaack/ml/basics/config/properties/ConfigReader.java
@@ -3,10 +3,13 @@ package de.schaack.ml.basics.config.properties;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
+import java.util.logging.Logger;
+
 
 public class ConfigReader {
     private static final String[] CONFIG_FILES = { "application.yaml", "application.yml", "application.properties" };
     private static Properties properties;
+    private static Logger globalLogger = Logger.getGlobal();
 
     private ConfigReader() {
     }
@@ -16,7 +19,7 @@ public class ConfigReader {
     }
 
     public static void loadConfig() {
-        System.out.println("Start loading configuration");
+        globalLogger.info("Start loading configuration");
         for (String fileName : CONFIG_FILES) {
             try (InputStream input = ConfigReader.class.getClassLoader().getResourceAsStream(fileName)) {
                 if (input != null) {
@@ -25,11 +28,11 @@ public class ConfigReader {
                     } else if (fileName.endsWith(".properties")) {
                         properties.load(input);
                     }
-                    System.out.println("Loaded configuration from: " + fileName);
+                    globalLogger.info("Loaded configuration from: " + fileName);
                     return; // Exit after successfully loading a file
                 }
             } catch (IOException ex) {
-                System.err.println("Error loading " + fileName + ": " + ex.getMessage());
+                globalLogger.severe("Error loading " + fileName + ": " + ex.getMessage());
             }
         }
         throw new RuntimeException("No valid configuration file found");
@@ -39,7 +42,7 @@ public class ConfigReader {
         return properties.getProperty(key);
     }
 
-    public static Properties getProperties(){
+    public static Properties getProperties() {
         return properties;
     }
 }