+++ src/main/java/de/schaack/ml/basics/functions/activation/ActivationFunction.java
@@ -1,0 +1,7 @@
+package de.schaack.ml.basics.functions.activation;
+
+
+public interface ActivationFunction {
+   double activate(double x);
+}
+