+++ src/main/java/de/schaack/ml/basics/functions/activation/ActivationFunctionInterface.java
@@ -1,6 +1,0 @@
-package de.schaack.ml.basics.functions.activation;
-
-public interface ActivationFunctionInterface {
-   
-}
-