+++ src/main/java/de/schaack/ml/basics/functions/evaluation/classification/binary/BinaryConfusionMatrix.java
@@ -1,9 +1,9 @@
-package de.schaack.ml.basics.functions.evaluations.classification.binary;
+package de.schaack.ml.basics.functions.evaluation.classification.binary;
 
-import static de.schaack.ml.basics.functions.evaluations.classification.binary.BinaryEvaluationEnum.FALSE_NEGATIVE;
-import static de.schaack.ml.basics.functions.evaluations.classification.binary.BinaryEvaluationEnum.FALSE_POSITIVE;
-import static de.schaack.ml.basics.functions.evaluations.classification.binary.BinaryEvaluationEnum.TRUE_NEGATIVE;
-import static de.schaack.ml.basics.functions.evaluations.classification.binary.BinaryEvaluationEnum.TRUE_POSITIVE;
+import static de.schaack.ml.basics.functions.evaluation.classification.binary.BinaryEvaluationEnum.FALSE_NEGATIVE;
+import static de.schaack.ml.basics.functions.evaluation.classification.binary.BinaryEvaluationEnum.FALSE_POSITIVE;
+import static de.schaack.ml.basics.functions.evaluation.classification.binary.BinaryEvaluationEnum.TRUE_NEGATIVE;
+import static de.schaack.ml.basics.functions.evaluation.classification.binary.BinaryEvaluationEnum.TRUE_POSITIVE;
 
 import java.util.Collection;
 import java.util.EnumMap;
@@ -11,7 +11,7 @@ import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.StreamSupport;
 
-import de.schaack.ml.basics.config.implementations.BinaryConfusionMatrixConfig;
+import de.schaack.ml.basics.models.settings.impl.BinaryConfusionMatrixConfig;
 
 public class BinaryConfusionMatrix {