+++ src/main/java/de/schaack/ml/basics/functions/evaluation/classification/binary/BinaryEvaluationEnum.java
@@ -1,4 +1,4 @@
-package de.schaack.ml.basics.functions.evaluations.classification.binary;
+package de.schaack.ml.basics.functions.evaluation.classification.binary;
 
 public enum BinaryEvaluationEnum {
     TRUE_POSITIVE, TRUE_NEGATIVE, FALSE_POSITIVE, FALSE_NEGATIVE