+++ src/main/java/de/schaack/ml/basics/functions/loss/LossFunction.java
@@ -1,5 +1,5 @@
 package de.schaack.ml.basics.functions.loss;
 
-public interface LossFunctionInterface {
+public interface LossFunction {
     
 }