+++ src/main/java/de/schaack/ml/basics/functions/optimizer/OptimizerFunction.java
@@ -1,0 +1,5 @@
+package de.schaack.ml.basics.functions.optimizer;
+
+public interface OptimizerFunction {
+    
+}