+++ src/main/java/de/schaack/ml/basics/models/Model.java
@@ -1,5 +1,0 @@
-package de.schaack.ml.basics.models;
-
-public abstract class Model {
-  
-}