+++ src/main/java/de/schaack/ml/basics/models/classification/Perceptron.java
@@ -1,7 +1,0 @@
-package de.schaack.ml.basics.models.classification;
-
-import de.schaack.ml.basics.models.Model;
-
-public class Perceptron extends Model {
-    
-}