+++ src/main/java/de/schaack/ml/basics/models/interfaces/ModelInterface.java
@@ -1,0 +1,8 @@
+package de.schaack.ml.basics.models.interfaces;
+
+import de.schaack.ml.basics.data.DataSet;
+
+public interface ModelInterface {    
+    void train(DataSet trainingData);
+
+}