+++ src/main/java/de/schaack/ml/basics/models/interfaces/Regressor.java
@@ -1,0 +1,7 @@
+package de.schaack.ml.basics.models.interfaces;
+
+import de.schaack.ml.basics.data.DataRow;
+
+public interface Regressor extends ModelInterface {
+    double predict(DataRow point);
+}