+++ src/main/java/de/schaack/ml/basics/models/settings/ModelSettings.java
@@ -1,0 +1,7 @@
+package de.schaack.ml.basics.models.settings;
+
+import java.util.Map;
+
+public interface ModelSettings {
+    Map<String, Object> getParameters();
+}