+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] 2024-07-20
+### ADDED
+- Model Interfaces and example implementation with a Perceptron
+    - abstract BaseModel with:
+        - ModelInterface
+            - Classifier
+            - Clusterer
+            - Regressor
+        - ModelSettings
+### CHANGED
+- Refactored Logging
+    - Now injectable into Models
+    - Default stays as previous
+- Renamed packages
+
 ## [0.0.3] 2024-07-20
 ### ADDED
 - Basic buffered Logger setup